requirements.txt
requests==2.31.0
moviepy==1.0.3
pillow==9.2.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-api-python-client==2.98.0
imageio-ffmpeg==0.6.0
playwright==1.52.0
huggingface_hub==0.34.4
supriya==26.3b0
